+++ src/blob/header.rs
@@ -30,8 +30,8 @@ impl Header {
             .await
             .with_context(|| format!("failed to open blob file: {}", name))?;
         let size = serialized_size(&Header::new()).expect("failed to serialize default header");
-        let mut buf = vec![0; size as usize];
-        file.read_at(&mut buf, 0)
+        let buf = file
+            .read_exact_at_allocate(size as usize, 0)
             .await
             .with_context(|| format!("failed to read from file: {}", name))?;
         let header: Self = deserialize(&buf)
@@ -59,5 +59,10 @@ impl Header {
             return Err(Error::validation(param, "blob header magic byte mismatch").into());
         }
         Ok(())
+    }
+
+    #[inline]
+    pub(crate) fn serialized_size(&self) -> u64 {
+        serialized_size(&self).expect("blob header size")
     }
 }