+++ CHANGELOG.md
@@ -4,11 +4,55 @@ Pearl changelog
 
 ## [Unreleased]
 #### Added
-- Add functions to get info about blobs and indexes (#151)
-- Add key_size field to index header (#151)
 
 
 #### Changed
+
+
+#### Fixed
+
+
+#### Updated
+
+
+
+## [0.17.0] - 2023-01-16
+#### Added
+- Special handling for large data write (#192)
+
+#### Changed
+- Added different path for sync read/write (#184)
+
+#### Fixed
+- Fixed skip condition in Storage::get_data_last causing incorrect NotFound result (#235)
+
+#### Updated
+
+
+
+## [0.16.0] - 2022-12-30
+#### Added
+- Corrupted blobs count calculation (#224)
+
+#### Changed
+- Return delete or not found as distinct results (#208)
+- Save delete records in index (#207)
+- Use bytes for arguments (#191)
+
+#### Fixed
+
+
+#### Updated
+- Crate `crc` (#225)
+
+
+## [0.15.0] - 2022-11-22
+#### Added
+- Add functions to get info about blobs and indexes (#151)
+- Add key_size field to index header (#151)
+
+#### Changed
+- Change file operations to work on full buffers (#190)
 - Remove iterators from bloom filter (#194)
 - Excessive key conversion removed (#193)
 - Removed serialization from critical section (#188)